Move ConcurrentMockTracer to wisp-tracing-testing
