+++ settings.gradle.kts
@@ -33,3 +33,4 @@ include(":wisp-time-testing")
 include(":wisp-token")
 include(":wisp-token-testing")
 include(":wisp-tracing")
+include(":wisp-tracing-testing")
