+++ wisp-tracing-testing/README.md
@@ -1,0 +1,5 @@
+# wisp-tracing-testing
+
+Testing utilities for wisp-tracing.
+
+Provides an alternative to `io.opentracing.MackTracer`, which is safe for concurrent test environments.
