+++ wisp-tracing-testing/build.gradle.kts
@@ -1,0 +1,8 @@
+plugins {
+    `java-library`
+}
+
+dependencies {
+    api(libs.openTracing)
+    implementation(libs.openTracingMock)
+}
