+++ wisp-tracing-testing/gradle.properties
@@ -1,0 +1,4 @@
+POM_ARTIFACT_ID=wisp-tracing-testing
+POM_NAME=wisp-tracing-testing
+POM_DESCRIPTION=concurrency-safe test helpers for tracing
+POM_PACKAGING=jar
